+++ src/runtime/server/index.ts
@@ -1,2 +1,1 @@
-export { createPDF } from './pdfkit'
-export { streamReturnPDF } from './streamReturn'
+export { createPDF, streamReturnPDF } from './pdf'
