+++ src/runtime/server/pdf.ts
@@ -1,5 +1,7 @@
-import type { WriteStream } from 'node:fs'
 import PDFDocument from 'pdfkit'
+import { sendStream, setHeader } from 'h3'
+import type { H3Event } from 'h3'
+import type { WriteStream } from 'node:fs'
 
 type PDFDocumentType<TData> = typeof PDFDocument & {
   data?: TData
@@ -38,3 +40,26 @@ export function createPDF<TData>(options: PDFKit.PDFDocumentOptions = defaultPDF
 
   return doc
 }
+
+/**
+ * Render a PDFDocument using an H3Event.
+ * Return this function at the end of your event handler.
+ *
+ * @param event The H3Event passed from the Event handler
+ * @param pdf The created PDF Object from createPDF. This can be a promise or a static object
+ */
+export const streamReturnPDF = async (event: H3Event, pdf: PDFKit.PDFDocument | Promise<PDFKit.PDFDocument>) => {
+  try {
+    const doc = await Promise.resolve(pdf)
+    if (doc.info.Title) {
+      setHeader(event, 'Content-disposition', `filename=${doc.info.Title}.pdf`)
+    }
+
+    setHeader(event, 'Content-Type', 'application/pdf')
+    return sendStream(event, doc as any)
+  } catch(error) {
+    console.error('nuxt-pdf: Error during PDF generation:', error)
+    throw new Error('PDF generated failed.')
+  }
+}
+
